fix(rag-eval): _cosine leaked numpy scalars, breaking response serialization
_cosine's `dot` stayed a numpy.float32/64 even after norm1/norm2 were
coerced to plain Python floats via math.sqrt (numpy dtype propagates
through the a*b/sum arithmetic on numpy-array inputs). That numpy scalar
ended up inside RagEvalResponse.details, and Pydantic's JSON serializer
rejects numpy types outright:
PydanticSerializationError: Unable to serialize unknown type: <class 'numpy.float32'>.

This 500'd at FastAPI's response-serialization stage — after the route
handler's own try/except had already returned "successfully" — so it was
invisible to that handler's error mapping and showed up as a bare 500
"Internal Server Error" to the caller. Caught live, immediately after the
prior `if not v1` truthiness fix, testing the real qwen3-embedding-4b path
end-to-end (both bugs only manifest with real numpy-array embeddings; every
existing fake adapter in tests returned plain Python lists).

Cast to float() at the end of _cosine. Added a unit test asserting the
return type, plus a router-level integration test using a numpy-shaped
fake adapter (matching UnifiedEmbeddingAdapter.encode()'s real return
shape) to catch this class of bug end-to-end, not just at the unit level.
774/774 passing.

+++ analysis/rag_eval.py
@@ -71,7 +71,13 @@ def _cosine(v1: list, v2: list) -> float:
     norm2 = math.sqrt(sum(b * b for b in v2))
     if norm1 == 0 or norm2 == 0:
         return 0.0
-    return dot / (norm1 * norm2)
+    # v1/v2 elements are numpy.float32/64 for a real embedding adapter's
+    # output, and numpy scalar dtypes propagate through the arithmetic above
+    # (dot stays a numpy scalar even though norm1/norm2 are plain Python
+    # floats via math.sqrt) — cast back so this always returns a plain
+    # Python float. Pydantic can't JSON-serialize numpy scalar types, and
+    # this value ends up embedded in RagEvalResponse.details.
+    return float(dot / (norm1 * norm2))
 
 
 def compute_context_precision(question: str, contexts: list, *, embed_fn=None) -> dict:
